Feat: separate DeeplinkManager.init()and .scan() for repeated invocations
- Move ScannedBike to its own file
